backend/pyproject.toml
[project]
name = "ncpi-concept-search"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai-slim[anthropic]",
    "pydantic-evals",
    "python-dotenv",
]

[project.scripts]
concept-search = "concept_search.cli:main"
